+++ mediation/Pangle/source/plugin/Assets/GoogleMobileAds/Mediation/Pangle/Platforms/iOS/PangleClient.cs
@@ -12,21 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#if UNITY_ANDROID
+#if UNITY_IOS
 
 using UnityEngine;
 
-using GoogleMobileAds.Common.Mediation.Pangle;
+using GoogleMobileAds.Mediation.Pangle.Common;
 
-namespace GoogleMobileAds.Android.Mediation.Pangle
+namespace GoogleMobileAds.Mediation.Pangle.iOS
 {
     public class PangleClient : IPangleClient
     {
         private static PangleClient instance = new PangleClient();
         private PangleClient() { }
-
-        private const string pangleMediationAdapterClassName =
-               "com.google.ads.mediation.pangle.PangleMediationAdapter";
 
         public static PangleClient Instance
         {
@@ -38,16 +35,12 @@ namespace GoogleMobileAds.Android.Mediation.Pangle
 
         public void SetGDPRConsent(int gdpr)
         {
-            AndroidJavaClass PangleMediationAdapter =
-                    new AndroidJavaClass(pangleMediationAdapterClassName);
-            PangleMediationAdapter.CallStatic("setGDPRConsent", gdpr);
+            Externs.GADUMPangleSetGDPRConsent(gdpr);
         }
 
         public void SetDoNotSell(int doNotSell)
         {
-            AndroidJavaClass PangleMediationAdapter =
-                    new AndroidJavaClass(pangleMediationAdapterClassName);
-            PangleMediationAdapter.CallStatic("setDoNotSell", doNotSell);
+            Externs.GADUMPangleSetDoNotSell(doNotSell);
         }
     }
 }